feat: actions sur les alarmes depuis la table de résultat

+++ ui/src/views/GetVision/alarms/ParcAlarmsTab/AlarmsTable/AlarmsActions.vue
@@ -20,9 +20,13 @@ export default {
       let additionalActions = [];
 
       if (this.alarm.disabled) {
-        additionalActions.push('actions.DISABLE');
-      } else {
         additionalActions.push('actions.ENABLE');
+      } else {
+        additionalActions.push('actions.DISABLE');
+      }
+
+      if (!this.alarm.numberOfTargetedLines) {
+        additionalActions.push('actions.DELETE');
       }
 
       return ['getsim.actions.DETAIL', ...additionalActions];
@@ -36,9 +40,19 @@ export default {
   },
 
   methods: {
-    ...mapMutations(['flashMessage']),
+    ...mapMutations(['flashMessage', 'confirmAction']),
 
     async onActionClicked(action) {
+      const showMessage = ret => {
+        if (ret) {
+          this.flashMessage({ level: 'success', message: this.$t('genericSuccessMessage') });
+        } else {
+          this.flashMessage({ level: 'danger', message: this.$t('genericErrorMessage') });
+        }
+
+        this.$emit('actionIsDone');
+      };
+
       if (action === 'getsim.actions.DETAIL') {
         this.$router.push({ name: 'alarmDetail', params: { alarmId: this.alarm.id } });
       }
@@ -46,27 +60,34 @@ export default {
       if (!this.canDoAsyncAction) return;
 
       let response;
-      this.canDoAsyncAction = false;
 
       if (action === 'actions.ENABLE') {
+        this.canDoAsyncAction = false;
         response = await enableAlarm(this.alarm.id);
+        this.canDoAsyncAction = true;
       }
 
       if (action === 'actions.DISABLE') {
+        this.canDoAsyncAction = false;
         response = await disableAlarm(this.alarm.id);
+        this.canDoAsyncAction = true;
       }
 
       if (action === 'actions.DELETE') {
-        response = await deleteAlarm(this.alarm.id);
+        this.confirmAction({
+          message: 'confirmAction',
+          actionFn: async () => {
+            this.canDoAsyncAction = false;
+            response = await deleteAlarm(this.alarm.id);
+            this.canDoAsyncAction = true;
+
+            showMessage(response);
+          },
+        });
+        return;
       }
 
-      this.canDoAsyncAction = true;
-
-      if (response) {
-        this.flashMessage({ level: 'success', message: this.$t('genericSuccessMessage') });
-      } else {
-        this.flashMessage({ level: 'danger', message: this.$t('genericErrorMessage') });
-      }
+      showMessage(response);
     },
   },
 };
